pyproject.toml
[tool.poetry]
name = "telefront"
version = "0.1.0"
description = ""
authors = ["cabup <pass@localhost>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.migrator.dependencies]
python = "^3.12"
jinja2 = "^3.1.6"
redis = "^5.2.1"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"

[tool.poetry.group.bot.dependencies]
python = "^3.12"
jinja2 = "^3.1.6"
redis = "^5.2.1"
sqlalchemy = "^2.0.40"
flask = "^3.1.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
